+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/jira/StandaloneFormula.kt
@@ -47,7 +47,8 @@ class StandaloneFormula private constructor(
     private val overriddenNetwork: Network? = null,
     private val databaseComputer: Computer,
     private val databaseVolume: Volume,
-    private val accessRequester: AccessRequester
+    private val accessRequester: AccessRequester,
+    private val adminPasswordPlainText: String
 ) : JiraFormula {
     private val logger: Logger = LogManager.getLogger(this::class.java)
 
@@ -75,7 +76,8 @@ class StandaloneFormula private constructor(
         stackCreationTimeout = Duration.ofMinutes(30),
         databaseComputer = M4ExtraLargeElastic(),
         databaseVolume = Volume(100),
-        accessRequester = Defaults.accessRequester
+        accessRequester = Defaults.accessRequester,
+        adminPasswordPlainText = "admin"
     )
 
     @Suppress("DEPRECATION")
@@ -96,7 +98,8 @@ class StandaloneFormula private constructor(
         stackCreationTimeout = Duration.ofMinutes(30),
         databaseComputer = M4ExtraLargeElastic(),
         databaseVolume = Volume(100),
-        accessRequester = Defaults.accessRequester
+        accessRequester = Defaults.accessRequester,
+        adminPasswordPlainText = "admin"
     )
 
     override fun provision(
@@ -187,7 +190,8 @@ class StandaloneFormula private constructor(
             databaseIp = databasePrivateIp,
             productDistribution = productDistribution,
             ssh = jiraSsh,
-            computer = computer
+            computer = computer,
+            adminPasswordPlainText  = adminPasswordPlainText
         )
 
         val jiraNodeSecurityGroup = jiraStack.findSecurityGroup("JiraNodeSecurityGroup")
@@ -229,12 +233,12 @@ class StandaloneFormula private constructor(
                 databaseComputer.setUp(it)
                 logger.info("Setting up database...")
                 key.get().file.facilitateSsh(databaseSshIp)
-                val location = database.setup(it)
+                val databaseDataLocation = database.setup(it)
                 logger.info("Database is set up")
                 logger.info("Starting database...")
                 database.start(jiraPublicHttpAddress, it)
                 logger.info("Database is started")
-                RemoteLocation(databaseSsh.host, location)
+                RemoteLocation(databaseSsh.host, databaseDataLocation)
             }
         }
 
@@ -294,6 +298,7 @@ class StandaloneFormula private constructor(
         private var databaseComputer: Computer = M4ExtraLargeElastic()
         private var databaseVolume: Volume = Volume(100)
         private var accessRequester: AccessRequester = Defaults.accessRequester
+        private var adminPasswordPlainText: String = "admin"
 
         internal constructor(
             formula: StandaloneFormula
@@ -310,6 +315,7 @@ class StandaloneFormula private constructor(
             network = formula.overriddenNetwork
             databaseComputer = formula.databaseComputer
             databaseVolume = formula.databaseVolume
+            adminPasswordPlainText = formula.adminPasswordPlainText
         }
 
         fun config(config: JiraNodeConfig): Builder = apply { this.config = config }
@@ -321,6 +327,8 @@ class StandaloneFormula private constructor(
 
         fun databaseComputer(databaseComputer: Computer): Builder = apply { this.databaseComputer = databaseComputer }
         fun databaseVolume(databaseVolume: Volume): Builder = apply { this.databaseVolume = databaseVolume }
+
+        fun adminPasswordPlainText(adminPasswordPlainText: String): Builder = apply { this.adminPasswordPlainText = adminPasswordPlainText }
 
         internal fun network(network: Network) = apply { this.network = network }
 
@@ -338,7 +346,8 @@ class StandaloneFormula private constructor(
             overriddenNetwork = network,
             databaseComputer = databaseComputer,
             databaseVolume = databaseVolume,
-            accessRequester = accessRequester
+            accessRequester = accessRequester,
+            adminPasswordPlainText  = adminPasswordPlainText
         )
     }
 }